refactor: rename UserModel to UserDto and refactor related files
