+++ chat/src/main/kotlin/net/thechance/chat/api/dto/UserDto.kt
@@ -1,16 +1,16 @@
-package net.thechance.chat.service.model
+package net.thechance.chat.api.dto
 
 import net.thechance.chat.entity.ContactUser
 
-data class UserModel(
+data class UserDto(
     val firstName: String,
     val lastName: String,
     val phoneNumber: String,
     val imageUrl: String? = null,
 )
 
-fun ContactUser.toModel(): UserModel{
-    return UserModel(
+fun ContactUser.toDto(): UserDto{
+    return UserDto(
         firstName = this.firstName,
         lastName = this.lastName,
         phoneNumber = this.phoneNumber,
